pc_program/requirements.txt
# 누룽지 생산량 카운팅 시스템 - PC 프로그램 의존성

# MQTT 통신
paho-mqtt>=1.6.1

# 이미지 처리 (옵션)
pillow>=10.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# GUI (Python 기본 포함, 별도 설치 불필요)
# tkinter

# 유틸리티
python-dateutil>=2.8.2
